requirements.txt
pypinyin>=0.49.0
rapidfuzz>=3.0.0
dbus-python>=1.3.2
PyQt5>=5.15.0
